\name{describe_cal}
\alias{describe_cal}
\docType{package}
\title{
Compute summary statistics for meetings
}
\arguments{
  \item{df}{The anpai dataframe}
  \item{start_date}{Optional, Use Y-m-d to specify}
  \item{end_date}{Optional, Use Y-m-d to specify}
}
\description{
Computes summary statistics for the meetings
}
\usage{
describe_cal(df, start_date, end_date)
}
\examples{
\dontrun{
describe_cal(df, 2022-01-01, 2022-01-12)
}
}
