\name{plot_breaks}
\alias{plot_breaks}
\docType{package}
\title{
Plot median break times for next week
}
\arguments{
  \item{df}{The anpai dataframe}
}
\description{
Plot median break times for next week and indicate what types of tasks are feasible within the specific breaks
}
\usage{
plot_breaks(df)
}
\examples{
\dontrun{
plot_breaks(df)
}
}
